Add page types attribute assignment handling
