+++ src/components/AssignAttributeDialog/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./AssignAttributeDialog";
+export * from "./AssignAttributeDialog";
