+++ src/components/AttributeUnassignDialog/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./AttributeUnassignDialog";
+export * from "./AttributeUnassignDialog";
