+++ src/components/BulkAttributeUnassignDialog/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./BulkAttributeUnassignDialog";
+export * from "./BulkAttributeUnassignDialog";
