+++ src/fragments/types/AvailableAttributeFragment.ts
@@ -1,0 +1,14 @@
+/* tslint:disable */
+/* eslint-disable */
+// This file was automatically generated and should not be edited.
+
+// ====================================================
+// GraphQL fragment: AvailableAttributeFragment
+// ====================================================
+
+export interface AvailableAttributeFragment {
+  __typename: "Attribute";
+  id: string;
+  name: string | null;
+  slug: string | null;
+}
