+++ src/fragments/types/PageTypeDetailsFragment.ts
@@ -31,34 +31,6 @@ export interface PageTypeDetailsFragment_attributes {
   filterableInStorefront: boolean;
 }
 
-export interface PageTypeDetailsFragment_availableAttributes_pageInfo {
-  __typename: "PageInfo";
-  hasNextPage: boolean;
-  endCursor: string | null;
-}
-
-export interface PageTypeDetailsFragment_availableAttributes_edges_node {
-  __typename: "Attribute";
-  id: string;
-  name: string | null;
-  slug: string | null;
-  type: AttributeTypeEnum | null;
-  visibleInStorefront: boolean;
-  filterableInDashboard: boolean;
-  filterableInStorefront: boolean;
-}
-
-export interface PageTypeDetailsFragment_availableAttributes_edges {
-  __typename: "AttributeCountableEdge";
-  node: PageTypeDetailsFragment_availableAttributes_edges_node;
-}
-
-export interface PageTypeDetailsFragment_availableAttributes {
-  __typename: "AttributeCountableConnection";
-  pageInfo: PageTypeDetailsFragment_availableAttributes_pageInfo;
-  edges: PageTypeDetailsFragment_availableAttributes_edges[];
-}
-
 export interface PageTypeDetailsFragment {
   __typename: "PageType";
   id: string;
@@ -66,5 +38,4 @@ export interface PageTypeDetailsFragment {
   metadata: (PageTypeDetailsFragment_metadata | null)[];
   privateMetadata: (PageTypeDetailsFragment_privateMetadata | null)[];
   attributes: (PageTypeDetailsFragment_attributes | null)[] | null;
-  availableAttributes: PageTypeDetailsFragment_availableAttributes | null;
 }
