+++ src/pageTypes/types/AssignPageAttribute.ts
@@ -37,34 +37,6 @@ export interface AssignPageAttribute_pageAttributeAssign_pageType_attributes {
   filterableInStorefront: boolean;
 }
 
-export interface AssignPageAttribute_pageAttributeAssign_pageType_availableAttributes_pageInfo {
-  __typename: "PageInfo";
-  hasNextPage: boolean;
-  endCursor: string | null;
-}
-
-export interface AssignPageAttribute_pageAttributeAssign_pageType_availableAttributes_edges_node {
-  __typename: "Attribute";
-  id: string;
-  name: string | null;
-  slug: string | null;
-  type: AttributeTypeEnum | null;
-  visibleInStorefront: boolean;
-  filterableInDashboard: boolean;
-  filterableInStorefront: boolean;
-}
-
-export interface AssignPageAttribute_pageAttributeAssign_pageType_availableAttributes_edges {
-  __typename: "AttributeCountableEdge";
-  node: AssignPageAttribute_pageAttributeAssign_pageType_availableAttributes_edges_node;
-}
-
-export interface AssignPageAttribute_pageAttributeAssign_pageType_availableAttributes {
-  __typename: "AttributeCountableConnection";
-  pageInfo: AssignPageAttribute_pageAttributeAssign_pageType_availableAttributes_pageInfo;
-  edges: AssignPageAttribute_pageAttributeAssign_pageType_availableAttributes_edges[];
-}
-
 export interface AssignPageAttribute_pageAttributeAssign_pageType {
   __typename: "PageType";
   id: string;
@@ -72,7 +44,6 @@ export interface AssignPageAttribute_pageAttributeAssign_pageType {
   metadata: (AssignPageAttribute_pageAttributeAssign_pageType_metadata | null)[];
   privateMetadata: (AssignPageAttribute_pageAttributeAssign_pageType_privateMetadata | null)[];
   attributes: (AssignPageAttribute_pageAttributeAssign_pageType_attributes | null)[] | null;
-  availableAttributes: AssignPageAttribute_pageAttributeAssign_pageType_availableAttributes | null;
 }
 
 export interface AssignPageAttribute_pageAttributeAssign {
