--- src/productTypes/components/AssignAttributeDialog/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./AssignAttributeDialog";
-export * from "./AssignAttributeDialog";
