--- src/productTypes/components/ProductTypeAttributeUnassignDialog/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ProductTypeAttributeUnassignDialog";
-export * from "./ProductTypeAttributeUnassignDialog";
