--- src/productTypes/components/ProductTypeBulkAttributeUnassignDialog/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ProductTypeBulkAttributeUnassignDialog";
-export * from "./ProductTypeBulkAttributeUnassignDialog";
